[#175717757] Fixes cut content when hiding footer button
