+++ ts/features/bonus/bpd/screens/details/BpdDetailsScreen.tsx
@@ -16,6 +16,11 @@ import { bpdUnsubscriptionSelector } from "../../store/reducers/details/activati
 import BpdPeriodSelector from "./BpdPeriodSelector";
 import BpdPeriodDetail from "./periods/BpdPeriodDetail";
 import GoToTransactions from "./transaction/GoToTransactions";
+import { fromNullable } from "fp-ts/lib/Option";
+import * as pot from "italia-ts-commons/lib/pot";
+import { bpdTransactionsForSelectedPeriod } from "../../store/reducers/details/transactions";
+import { bpdSelectedPeriodSelector } from "../../store/reducers/details/selectedPeriod";
+import { navigateToBpdTransactions } from "../../navigation/actions";
 
 export type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps>;
@@ -54,6 +59,26 @@ const BpdDetailsScreen: React.FunctionComponent<Props> = props => {
     }
   }, [props.unsubscription]);
 
+  /**
+   * Display the transactions button when:
+   * - Period is closed and transactions number is > 0
+   * - Period is active
+   * never displays for inactive/incoming period
+   */
+  const canRenderButton = fromNullable(props.selectedPeriod).fold(false, sp => {
+    switch (sp.status) {
+      case "Closed":
+        return pot.getOrElse(
+          pot.map(props.transactions, val => val.length > 0),
+          false
+        );
+      case "Inactive":
+        return false;
+      default:
+        return true;
+    }
+  });
+
   return (
     <LoadingSpinnerOverlay
       isLoading={loading}
@@ -68,7 +93,11 @@ const BpdDetailsScreen: React.FunctionComponent<Props> = props => {
         topContent={<View style={styles.headerSpacer} />}
         gradientHeader={true}
         hideHeader={true}
-        footerContent={<GoToTransactions />}
+        footerContent={
+          canRenderButton && (
+            <GoToTransactions goToTransactions={props.goToTransactions} />
+          )
+        }
       >
         <View style={styles.selector}>
           <BpdPeriodSelector />
@@ -85,11 +114,14 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   completeUnsubscription: () => {
     dispatch(bpdUnsubscribeCompleted());
     dispatch(NavigationActions.back());
-  }
+  },
+  goToTransactions: () => dispatch(navigateToBpdTransactions())
 });
 
 const mapStateToProps = (state: GlobalState) => ({
-  unsubscription: bpdUnsubscriptionSelector(state)
+  unsubscription: bpdUnsubscriptionSelector(state),
+  transactions: bpdTransactionsForSelectedPeriod(state),
+  selectedPeriod: bpdSelectedPeriodSelector(state)
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(BpdDetailsScreen);
